azure2/requirements.txt
torch
torchvision
opencv-python
numpy
scikit-learn
onnx
onnxruntime
tqdm
Pillow